Move dial addr logic to its own module
Dialing an address is not the concern of the orderbook. Create a
`dial_addr` submodule under the `http_api` module and move the code
over to it.
