+++ cnd/src/http_api/dial_addr.rs
@@ -1,0 +1,26 @@
+use crate::{facade::Facade, http_api::problem};
+use libp2p::Multiaddr;
+use serde::Deserialize;
+use warp::{Rejection, Reply};
+
+// TODO: Do we even want this module?
+
+#[derive(Deserialize, Debug)]
+pub struct DialPeerBody {
+    addresses: Vec<Multiaddr>,
+}
+
+pub async fn post_dial_addr(
+    body: serde_json::Value,
+    mut facade: Facade,
+) -> Result<impl Reply, Rejection> {
+    let body = DialPeerBody::deserialize(&body)
+        .map_err(anyhow::Error::new)
+        .map_err(problem::from_anyhow)
+        .map_err(warp::reject::custom)?;
+    // TODO: find out if the dial was successful?
+    for addr in body.addresses {
+        facade.dial_addr(addr).await;
+    }
+    Ok(warp::reply::reply())
+}
